WIP support for pre-install hooks
