+++ lib/cocoapods-fix-react-native/versions/0_54_2-post.rb
