+++ lib/cocoapods-fix-react-native/versions/0_54_4-post.rb
